Fix_bfillerror
Fixed an error on newer versions of Pandas where dataframe.reindex gave a dtype error. Work-around is somewhat clumsy - converted dataframe to a numpy array then back to a dataframe, followed by reindexing on the "time" level. Future versions of Pandas may make it so that a better solution is possible, and future iterations may incorporate that.