Workto20221014
Work up to 2022-10-14. Highlights:
Removed matrix method to discretize water volumes. Due to previous changes, this was giving a constant volume in each cell but still took time and large (100s of GB) amounts of ram for larger timeseries. Can be restored in the future if necessary, also turned into a loop to account for available RAM (this was then commented out).
Changes made either in this commit or the last one involved adding an explicit native soil compartment around the cell - this allowed for better estimation of the system's hydrology (for the Pine & 8th BC)